backend/requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
requests
python-dotenv
supabase
langchain
langchain-community
sentence-transformers
torch
